+++ src/settings.js
@@ -1255,8 +1255,8 @@ var WASM_BINARY_FILE = '';
 // For internal use only (name of the file containing asm.js, if relevant).
 var ASMJS_CODE_FILE = '';
 
-// For internal use only (name of the file containing pthread-main.js, if relevant).
-var PTHREAD_MAIN_FILE = '';
+// For internal use only (name of the file containing the pthread *.worker.js, if relevant).
+var PTHREAD_WORKER_FILE = '';
 
 // Base URL the source mapfile, if relevant
 var SOURCE_MAP_BASE = '';